hemeTools: simplify some packages
